Addressed comments by reviewer from pull request 288
