+++ scripts/postprocessing/coverage_analysis/activity_db_generator.py
@@ -1,4 +1,9 @@
 #!/usr/bin/python
+
+""" Creates a CSV file containing the Mapped Landmark features per robot pose published by
+    the /loc/ml/features topic
+"""
+
 import sys
 import rospy
 import numpy as np
@@ -24,12 +29,9 @@ class Activity_DBGenerator:
       if count == 0:
          robot_pose_str = "nan, nan, nan"
       else:
-         robot_pose = self.bodyFrame_from_navCam(data.pose.position.x, data.pose.position.y, data.pose.position.z,
+         robot_pose = self.body_frame_from_navcam(data.pose.position.x, data.pose.position.y, data.pose.position.z,
                 data.pose.orientation.x, data.pose.orientation.y, data.pose.orientation.z, data.pose.orientation.w)
 
-         #pose_str = "{:.3f} {:.3f} {:.3f} {:.3f} {:.3f} {:.3f} {:.3f}".format(
-         #data.pose.position.x, data.pose.position.y, data.pose.position.z,
-         #data.pose.orientation.x, data.pose.orientation.y, data.pose.orientation.z, data.pose.orientation.w)
          robot_pose_str = "{:.3f} {:.3f} {:.3f} {:.3f} {:.3f} {:.3f} {:.3f}".format(
                     robot_pose[0], robot_pose[1], robot_pose[2],
                     robot_pose[3], robot_pose[4], robot_pose[5], robot_pose[6])
@@ -40,7 +42,7 @@ class Activity_DBGenerator:
          ml_pose_str = "{:.3f} {:.3f} {:.3f}".format(data.landmarks[i].x, data.landmarks[i].y, data.landmarks[i].z)
          self.fileOut.write('{:s}\n'.format(ml_pose_str))
 
-   def bodyFrame_from_navCam(self, px,py,pz, qi,qj,qk,qw):
+   def body_frame_from_navcam(self, px,py,pz, qi,qj,qk,qw):
       # As defined in Astrobee repo: ${HOME}/astrobee/astrobee/config/robots/bumble.config
       pose_const = np.array([0.1157+0.002, -0.0422, -0.0826, 1.0])
       quat_const = [0.500, 0.500, 0.500, 0.500]
@@ -62,10 +64,10 @@ class Activity_DBGenerator:
 
       return [pose_out[0], pose_out[1], pose_out[2], quat_out[0], quat_out[1], quat_out[2], quat_out[3]]
 
-   def open_file(self, file, activity, date):
+   def open_file(self, file, activity, map, date):
       self.fileOut = open(file, "w")
-      self.fileOut.write("### Activity-Date\n### ML Number Robot Pose (x,y,z,r,p,y,w)\n### ML Coordinate(x,y,z)\n")
-      self.fileOut.write(activity + " " +date+ "\n")
+      self.fileOut.write("### Date-Map-Activity: " + date + "-" + map + "-" + activity +"\n")
+      self.fileOut.write("### ML Number Robot Pose (x,y,z,r,p,y,w)\n### ML Coordinate(x,y,z)\n")
 
    def close_file(self):
       self.fileOut.close()
@@ -79,19 +81,23 @@ if __name__ == '__main__':
 
    try:
 
-       if len(sys.argv) < 3:
-          print("Usage: activity_DBGenerator.py <activity date> <activity name>")
-          print("       <activity date> = 'YYYYMMDD'")
-          print("       <activity name> = 'activityName'")
+       if len(sys.argv) < 5:
+          print("Usage: activity_db_generator.py <activity date> <map name> <activity name> <location to save>")
+          print("       <activity date>    = 'YYYYMMDD'")
+          print("       <map name>         = 'mapName'")
+          print("       <activity name>    = 'activityName'")
+          print("       <location to save> = '/location/to/save/'")
           sys.exit(1)
 
        else:
           activity_date = sys.argv[1]
-          activity_name = sys.argv[2]
-          output_filename = activity_date + "_" + activity_name + "_db.csv"
+          map_name      = sys.argv[2]
+          activity_name = sys.argv[3]
+          location_name = sys.argv[4]
+          output_filename = location_name + activity_date + "_" + map_name + "_" + activity_name + "_db.csv"
 
        obj = Activity_DBGenerator()
-       obj.open_file(output_filename, activity_name, activity_date)
+       obj.open_file(output_filename, activity_name, map_name, activity_date)
        obj.listener()
 
    except KeyboardInterrupt:
